chore: minor improvements in front-end (#11)
* feat: added button to copy chunks from selected Collection

* fixed: unnecessary fetch calls in Dashboard Page

* feat: added the keyboard input options to handle chunks

* feat: notification for saving collection, using snackbar
